added a Boost class which stores solr query boost values. The class can be configured using the yacy.init file. The boost information is taken from the configuration each time when a query to solr is done.

+++ source/net/yacy/cora/document/analysis/EnhancedTextProfileSignature.java
@@ -57,7 +57,7 @@ import org.apache.solr.update.processor.Lookup3Signature;
 public class EnhancedTextProfileSignature extends Lookup3Signature {
 
   private float quantRate = 0.01f;
-  private float minTokenLen = 2;
+  private int   minTokenLen = 2;
   private StringBuilder evalText = new StringBuilder(120); // start with some capacity, makes it much faster.
 
   @Override
